backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.29
pydantic==2.6.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0